+++ main/Drivers/CAN/can_driver.h
@@ -7,6 +7,7 @@ extern "C" {
 #endif
 
 void canInit(void);
+void canReinit(void);
 twai_handle_t canGetBus0(void);
 twai_handle_t canGetBus1(void);
 void canResetBuses(void);