+++ main/Drivers/can_driver.h
@@ -8,6 +8,7 @@ extern "C" {
 #endif
 
 void canInit(void);
+void canReinit(void);
 twai_handle_t canGetBus0(void);   /* CAN1 */
 twai_handle_t canGetBus1(void);   /* CAN2 */